+++ spec/models/alchemy/page_spec.rb
@@ -2289,5 +2289,34 @@ module Alchemy
         end
       end
     end
+
+    describe '#fixed_attributes' do
+      let(:page) { Alchemy::Page.new }
+
+      it 'holds an instance of FixedAttributes' do
+        expect(page.fixed_attributes).to be_a(Alchemy::Page::FixedAttributes)
+      end
+    end
+
+    describe '#attribute_fixed?' do
+      let(:page) { Alchemy::Page.new }
+
+      it 'delegates to instance of FixedAttributes' do
+        expect_any_instance_of(Alchemy::Page::FixedAttributes).to receive(:fixed?).with('yolo')
+        page.attribute_fixed?('yolo')
+      end
+    end
+
+    describe '#set_fixed_attributes' do
+      context 'when fixed attributes are defined' do
+        let(:page) { create(:alchemy_page, page_layout: 'readonly') }
+
+        it 'sets them before each save' do
+          expect {
+            page.update(name: 'Foo')
+          }.to_not change { page.name }
+        end
+      end
+    end
   end
 end